Add a "cloneable" capability to repositories types.
Cloneable VCSs implement [versioncontrol_backend]_clone_path() and are able to
create a duplicate of a path of a repository and return a "possibly different"
repository and branch.

VCSs like Subversion will achieve this by creating a local copy within the same
repository and returning a different path, while decentralized VCSs will create
an entirely new repository with the copied contents of the initial repository.

Signed-off-by: Dan Hackney <dan@haxney.org>